testing & WIP

+++ yatc-t3-app/src/server/publish-tweet/tweet.drizzle.schema.ts
@@ -1,6 +1,6 @@
 import { relations } from "drizzle-orm";
 import { usersTable } from "../infrastructure/drizzle/base.drizzle.schema"
-import { integer, varchar, } from "drizzle-orm/pg-core";
+import { decimal, varchar } from "drizzle-orm/pg-core";
 import { drizzleTable } from "../infrastructure/drizzle/drizzleTable";
 
 export type TweetEntity = typeof tweetsTable.$inferSelect;
@@ -9,7 +9,7 @@ export const tweetsTable = drizzleTable("tweet", {
   id:           varchar("id", { length: 255 }).notNull().primaryKey(),
   text:         varchar("text", { length: 280 }).notNull(),
   publishedBy:  varchar("publishedBy", { length: 255 }).notNull(),
-  publishedAt:  integer("publishedAt").notNull(),
+  publishedAt:  decimal("publishedAt").notNull(),
 });
 
 export const tweetsTableRelations = relations(tweetsTable, ({one}) => ({
